requirements.txt
pillow==2.4.0
vokram==1.0.1
